check if file exist before saving

+++ Block11-upload-download-files/src/main/java/com/bosonit/garciajuanjo/Block11uploaddownloadfiles/services/impl/FileServiceImpl.java
@@ -1,11 +1,13 @@
 package com.bosonit.garciajuanjo.Block11uploaddownloadfiles.services.impl;
 
+import com.bosonit.garciajuanjo.Block11uploaddownloadfiles.exceptions.FileExistException;
 import com.bosonit.garciajuanjo.Block11uploaddownloadfiles.services.FileService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -21,18 +23,21 @@ public class FileServiceImpl implements FileService {
     private String pathFiles;
 
     @Override
-    public void store(MultipartFile file) throws IOException{
+    public void store(MultipartFile file) throws IOException, FileExistException {
         Path directory = Paths.get(pathFiles).toAbsolutePath();
+
+        File fileStore = new File(directory.toFile() + File.separator + file.getOriginalFilename());
+        if (fileStore.exists())
+            throw new FileExistException("The file " + file.getOriginalFilename() + " already exist in the directory " + directory);
 
         /*Con la función copy lo que hacemos es primero pasar el InputStream que vamos a copiar en la carpeta. Luego
         * con el segundo parámetro decimos que lo copiamos en el directorio (directory) con el nombre original que
         * nos viene en el MultipartFile*/
         Files.copy(file.getInputStream(), directory.resolve(Objects.requireNonNull(file.getOriginalFilename())));
-
     }
 
     @Override
-    public void store(List<MultipartFile> files) throws IOException{
+    public void store(List<MultipartFile> files) throws IOException, FileExistException {
         for (MultipartFile file :files) {
             this.store(file);
         }
